refactor(graperank): split schema types from service, merge service files
